requirements.txt
slack-bolt==1.18.1
openai[datalib]==1.35.0
azure-ai-inference==1.0.0b4
pyautogen==0.2.18
agents==0.1.5
python-dotenv==1.0.0
pydantic==2.7.4
asyncio==3.4.3
aiohttp==3.8.6
requests==2.31.0
mcp==1.0.0
httpx==0.27.0
google-api-python-client==2.114.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-auth==2.23.4
pytest==7.4.2
black==23.7.0
flake8==6.0.0 